==> 1_Prepare_Data.ipynb <==
### Общий план:
- __Заполнить пропуски в данных диагностики__
- __Агрегировать результаты диагностики по каждому юзеру__
- Выделить фичи из текста
- Агрегировать фичи по юзеру
- Объединить датасеты
- Построить матрицу корреляций по полученному датасету
- Наиболее удачные, яркие корреляции рассмотреть подробнее
- Построить бейзлайн, лр/бустинг, для прогнозирования результата диагностики по фичам текста

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

## Данные диагностик

train_df = []
test_df = []

for i in range(1,8):
    df = pd.read_excel('database0{}.xlsx'.format(i), header=1)
    train_df.append(df)

for i in range(8,10):
    df = pd.read_excel('database0{}.xlsx'.format(i), header=1)
    test_df.append(df)

df = pd.read_excel('database10.xlsx', header=1)    
test_df.append(df)    
    
train = pd.concat(train_df)
test = pd.concat(test_df)

train.shape, test.shape

train.columns

len(train['Unti ID'].unique()), len(test['Unti ID'].unique())

len(set(test['Unti ID']) - set(train['Unti ID']))

## Текст рефлексий

texts = pd.read_excel('Рефлексии_участников_задача_по_диагностике.xlsx')

texts.shape

len(texts['untiID'].unique())

texts.isna().sum()

train.shape, test.shape

len(train['Unti ID'].unique())

print("Кол-во уникальных ID в трейне, для к-х есть текстовые данные: {}".format(len(train['Unti ID'].unique()) - len(set(train['Unti ID'].unique()) - set(texts['untiID'].unique()))))
print("Кол-во уникальных ID в тесте, для к-х есть текстовые данные: {}".format(len(test['Unti ID'].unique()) - len(set(test['Unti ID'].unique()) - set(texts['untiID'].unique()))))



# ID-шники для которых нет текстов рефлексий
len(set(train['Unti ID'].unique()) - set(texts['untiID'].unique()))

# ID-шники для которых нет результатов диагностик
len(set(texts['untiID'].unique()) - set(train['Unti ID'].unique()))
# 450 unique text id - 244 = 206

# готовим трейн - берем id только тех у кого есть текстовые данные

train_ids = set(train['Unti ID']) - (set(train['Unti ID'].unique()) - set(texts['untiID'].unique()))
test_ids = set(test['Unti ID']) - (set(test['Unti ID'].unique()) - set(texts['untiID'].unique()))

df_train = train[train['Unti ID'].isin(train_ids)]
df_train.shape

df_test = test[test['Unti ID'].isin(test_ids)]
df_test.shape

### Заполнить пропуски в данных диагностики
Подход 1: заполнить нулем

len(df_train['Unti ID'].unique())

df_train[df_train['Unti ID']==103]

df_train.select_dtypes(include='object').isna().sum()

# convert to numeric
df_train = df_train.apply(pd.to_numeric, errors='coerce')
df_test = df_test.apply(pd.to_numeric, errors='coerce')

df_train_num = df_train.fillna(0)
df_test_num = df_test.fillna(0)

sum(df_train_num.isna().sum()), sum(df_test_num.isna().sum())

### Агрегировать результаты диагностики по каждому юзеру

from functools import reduce

def aggregate(data):
    sum_df = data.groupby(['Unti ID'], as_index=False).agg(np.sum)
    sum_df.columns = [str(col) + '_sum' if col != 'Unti ID' else 'Unti ID' for col in sum_df.columns]
    
    mean_df = data.groupby(['Unti ID'], as_index=False).agg(np.mean)
    mean_df.columns = [str(col) + '_mean' if col != 'Unti ID' else 'Unti ID' for col in mean_df.columns]
    
    median_df = data.groupby(['Unti ID'], as_index=False).agg(np.median)
    median_df.columns = [str(col) + '_median' if col != 'Unti ID' else 'Unti ID' for col in median_df.columns]
    
    max_df = data.groupby(['Unti ID'], as_index=False).agg(np.max)
    max_df.columns = [str(col) + '_max' if col != 'Unti ID' else 'Unti ID' for col in max_df.columns]
    
    min_df = data.groupby(['Unti ID'], as_index=False).agg(np.min)
    min_df.columns = [str(col) + '_min' if col != 'Unti ID' else 'Unti ID' for col in min_df.columns]
    
    data_frames = [sum_df, mean_df, median_df, max_df, min_df]
#     df_merged = pd.merge(sum_df, mean_df, on=['Unti ID'])
    df_merged = reduce(lambda  left,right: pd.merge(left,right,on=['Unti ID'], how='right'), data_frames)
    return df_merged

df_train_agg = aggregate(df_train_num)
df_train_agg.shape

# sanity check - all good
for col in df_train_agg.columns:
    if 'Большие данные (выходные)' in col:
        print(col)

df_test_agg = aggregate(df_test_num)
df_test_agg.shape

df_train_agg.rename(columns={'Unti ID':'untiID'}, inplace=True)
df_test_agg.rename(columns={'Unti ID':'untiID'}, inplace=True)

# Сохраним сформированные трейн и тест
df_train_agg.to_csv('diag_train_agg.csv',index=False)
df_test_agg.to_csv('diag_test_agg.csv',index=False)

### Фикс бага в данных рефлексий

# Странные id вопросов
texts['feedbackQuestionID'].unique()

texts[texts['feedbackQuestionID']=='Селина Алина Валерьевна'].head(3)

#### Похоже, что для некоторых сэмплов значения колонок сместились. Вернем id вопросов, вопросы и ответы в соответствующие колонки

ids_to_fix = []
for val in texts['feedbackQuestionID'].unique():
    if not isinstance(val, int):
        ids_to_fix.append(texts[texts['feedbackQuestionID']==val]['untiID'].unique()[0])

def fix_strange_ids(ids):
    for i in ids:
        texts['feedbackQuestionID'] = np.where(texts.untiID == i, texts['question'], texts['feedbackQuestionID'])
        texts['question'] = np.where(texts.untiID == i, texts['value'], texts['question'])
        texts['value'] = np.where(texts.untiID == i, texts['data'], texts['value'])
    return texts

texts.shape

text_df = fix_strange_ids(ids_to_fix)

# sanity check - all good
text_df['feedbackQuestionID'].unique()

text_df.shape, text_df.drop_duplicates().shape

# Убираем дубликаты
text_df.drop_duplicates(inplace=True)
text_df.shape

# Сохраним исправленные данные рефлексий
text_df.to_csv('fixed_reflex_texts.csv',index=False)
