==> 2_Additional_Data.ipynb <==
### Общий план:
- __Заполнить пропуски в данных диагностики__
- __Агрегировать результаты диагностики по каждому юзеру__
- Выделить фичи из текста
- Агрегировать фичи по юзеру
- Объединить датасеты
- Построить матрицу корреляций по полученному датасету
- Наиболее удачные, яркие корреляции рассмотреть подробнее
- Построить бейзлайн, лр/бустинг, для прогнозирования результата диагностики по фичам текста

import os
import numpy as np
import pandas as pd
import seaborn as sns
from functools import reduce
import matplotlib.pyplot as plt
%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

Работаем с дополнительными данными рефлексии, предоставленными организаторами.

# для воспроизведения - изменить путь!

path = r"\Дополнительные данные\Reflection_data+"

### Возьмем только анкеты с ответами на вопрос 'Чему Вы научились в рамках курса?'

add_data = []
for file in os.listdir(os.path.join(path)):
    df = pd.read_excel(os.path.join(path,file))
#     print(df.columns[1])
    if df.columns[1] == 'Чему Вы научились в рамках курса?' or df.columns[1] == 'Чему вы научились в рамках курса?':
        df.columns = ['untiID', 'Чему вы научились в рамках курса?']
        add_data.append(df)

# таких анкет - 51 из 69
len(add_data)

add_data = pd.concat(add_data)

add_data.shape

add_data.isna().sum()

add_data

### Объединим с данными диагностик

# для воспроизведения - изменить путь!
path_diag = r"\Дополнительные данные"

data_diag = pd.read_excel(os.path.join(path_diag, r'Диагностика_ПЦС_доп..xlsx'),header=1)
data_diag.shape

data_diag.rename(columns={"Leader ID": "untiID"}, inplace=True)

data_diag.columns

data = pd.merge(add_data ,data_diag, on=['untiID'])

### Заполнение пропусков в данных диагностики

data.dtypes

data.fillna(0, inplace=True)

sum(data.isna().sum())

def aggregate(data):
    sum_df = data.groupby(['untiID'], as_index=False).agg(np.sum)
    sum_df.columns = [str(col) + '_sum' if col != 'untiID' else 'untiID' for col in sum_df.columns]
    
    mean_df = data.groupby(['untiID'], as_index=False).agg(np.mean)
    mean_df.columns = [str(col) + '_mean' if col != 'untiID' else 'untiID' for col in mean_df.columns]
    
    median_df = data.groupby(['untiID'], as_index=False).agg(np.median)
    median_df.columns = [str(col) + '_median' if col != 'untiID' else 'untiID' for col in median_df.columns]
    
    max_df = data.groupby(['untiID'], as_index=False).agg(np.max)
    max_df.columns = [str(col) + '_max' if col != 'untiID' else 'untiID' for col in max_df.columns]
    
    min_df = data.groupby(['untiID'], as_index=False).agg(np.min)
    min_df.columns = [str(col) + '_min' if col != 'untiID' else 'untiID' for col in min_df.columns]
    
    data_frames = [sum_df, mean_df, median_df, max_df, min_df]
#     df_merged = pd.merge(sum_df, mean_df, on=['Unti ID'])
    df_merged = reduce(lambda  left,right: pd.merge(left,right,on=['untiID'], how='right'), data_frames)
    return df_merged

data_agg = aggregate(data.drop('Чему вы научились в рамках курса?',axis=1))

data_agg.shape

# Сохраним агрегированные диагностические доп.данные
data_agg.to_csv('additional_data_agg.csv',index=False)

# Сохраним агрегированные текстовые доп.данные
data[['untiID', 'Чему вы научились в рамках курса?']].to_csv('additional_data_text.csv',index=False)

### Объединим основной трейн (диагностики db0-7) и доп.данные

diag_train_agg = pd.read_csv('diag_train_agg.csv')

diag_train_agg.shape

# check if there are the same ids
print("Кол-во совпадающих ID:" , data_agg.shape[0] - len(set(data_agg.untiID) - set(diag_train_agg.untiID)))

# Определим совпадающие колонки
# Все колонки доп.данных есть в основном трейне - объединим датасеты по ним.
set(data_agg.columns) - set(diag_train_agg.columns)

diag_large_train = pd.concat([data_agg, diag_train_agg[data_agg.columns.tolist()]])
diag_large_train.shape

# sanity check
diag_large_train['творец_mean'].value_counts()

# сохраним трейн датасет с результатами диагностик
diag_large_train.to_csv('diag_large_train.csv',index=False)
