==> 3_Feature_Generation.ipynb <==
### Общий план:
- Заполнить пропуски в данных диагностики
- Агрегировать результаты диагностики по каждому юзеру
- __Выделить фичи из текста__
- __Агрегировать фичи по юзеру__
- __Объединить датасеты__
- Построить матрицу корреляций по полученному датасету
- Наиболее удачные, яркие корреляции рассмотреть подробнее
- Построить бейзлайн, лр/бустинг, для прогнозирования результата диагностики по фичам текста

import os
import numpy as np
import pandas as pd
import seaborn as sns
from functools import reduce
from deeppavlov import build_model, configs
import matplotlib.pyplot as plt
%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

# основные данные рефлексий
main_text = pd.read_csv('fixed_reflex_texts.csv')
main_text.shape

main_text['question'].unique()

# возьмем вопросы, для ответов на которые можно оценить тон учащегося
q = [
    'Чему новому, как вам кажется, вы научились?',
 'Была ли в этом ценность и в чем она выражается?',
 'Поделитесь вашими комментариями или пожеланиями по поводу того, с чем вы познакомились?',
 'Какие вопросы, комментарии  и пожелания у вас есть?',
 'Что вы поняли про свой проект по итогам общения с экспертами? К каким действиям это вас побудило?',
 'Что было полезно для вас лично и для команды в целом в трубе экспертов?',
 'А что можно было бы в трубе экспертов улучшить?',
 'Какие вопросы для вас остаются без ответа, чего вам не хватает сейчас?',
 'Если бы вы заранее знали, что труба будет проходить именно так, как бы вы к ней подготовились?',
    'Насколько ясно команда сформулировала, кому и с чем хочет помогать (проблема и пользователи)',
 'Насколько проработанный прототип у команды?',
 'Насколько решение отвечает проблеме?',
 'Уровень амбиций команды (насколько высокую планку поставили)',
 'Насколько хороша презентация команды (дизайн слайдов, структура, подача)',
 'Комментарии и рекомендации команде',
]

main_text = main_text[main_text['question'].isin(q)]
main_text.shape

# identify id where answers are not textual
ids_no_sentiment = []

for ind, val in zip(main_text.index, main_text['value']):
    if not isinstance(val, str):
        ids_no_sentiment.append(ind)
        
# slice rows where answers are textual
main_text = main_text[~main_text.index.isin(ids_no_sentiment)]
main_text.shape

# доп данные рефлексий
additional_data_text = pd.read_csv('additional_data_text.csv')
additional_data_text.shape

additional_data_text.dtypes

## Sentiment (Анализ тональности текста)

sent_model = build_model(configs.classifiers.rusentiment_elmo_twitter_cnn, download=True)

def do_sentiment(line):
    return sent_model([line])[0]

%%time
# Выделение тональности из основных данных
main_text['sentiment'] = main_text['value'].apply(do_sentiment)
main_text = pd.get_dummies(main_text, columns=['sentiment'])

%%time
# Выделение тональности из доп данных
additional_data_text['sentiment'] = additional_data_text['Чему вы научились в рамках курса?'].apply(do_sentiment)

additional_data_text['sentiment'].value_counts()

# encode sentiment as one-hot
additional_data_text = pd.get_dummies(additional_data_text, columns=['sentiment'])

# добавим колонки для сентиментов, найденных в основном датасете рефлексий
additional_data_text['sentiment_skip'] = 0
additional_data_text['sentiment_negative'] = 0

additional_data_text.head(3)

## NER (Именованные сущности)

ner_model = build_model(configs.ner.ner_rus, download=True)

def find_ner(line):
    words, tags = ner_model([line])
    try:
        for w, t in zip(words[0], tags[0]):
            if t != 'O':
                return 1
            else:
                return 0
    except:
        return 0

%%time
main_text['ner'] = main_text['value'].apply(find_ner)
additional_data_text['ner'] = additional_data_text['Чему вы научились в рамках курса?'].apply(find_ner)

additional_data_text['ner'].value_counts()

additional_data_text.head()

### Features: особенности текстов

import string
punctuation = string.punctuation

# длина комментария
main_text["answer_len"] = main_text['value'].apply(len)

# кол-во слов с заглавной буквы
main_text['upper_case_word_count'] = main_text['value'].apply(lambda x: len([wrd for wrd in x.split() if wrd[0].isupper()]))

# кол-во знаков препинания
main_text['punctuation_count'] = main_text['value'].apply(lambda x: len("".join(_ for _ in x if _ in punctuation)))

# длина комментария
additional_data_text["answer_len"] = additional_data_text['Чему вы научились в рамках курса?'].apply(len)

# кол-во слов с заглавной буквы
additional_data_text['upper_case_word_count'] = additional_data_text['Чему вы научились в рамках курса?'].apply(lambda x: len([wrd for wrd in x.split() if wrd[0].isupper()]))

# кол-во знаков препинания
additional_data_text['punctuation_count'] = additional_data_text['Чему вы научились в рамках курса?'].apply(lambda x: len("".join(_ for _ in x if _ in punctuation)))

### Features: использование частей речи

import pymorphy2
morph = pymorphy2.MorphAnalyzer()

def count_pos(line, pos):
    """
    Оставляет только верно написанные русские слова и подсчитываем,
    сколько слов определенной частей речи (аргумент pos) в сообщении.
    """
    char_regex = re.compile(r'[^а-яa-z]')
    line = char_regex.sub(' ', line.lower())
    
    chkr_ru = SpellChecker(d_ru)
    chkr_ru.set_text(line)
    mistakes_ru = set([err.word for err in chkr_ru])
    
    tokenized = toktok.tokenize(line)
    count = 0
    for i in tokenized:
        if i not in mistakes_ru:
            p = morph.parse(i)[0]
            if p.tag.POS==pos:
                count+=1
            
    return count

# подсчет существительных
main_text['nouns'] = main_text.apply(lambda x: count_pos(x['value'], 'NOUN'), axis=1)
additional_data_text['nouns'] = additional_data_text.apply(lambda x: count_pos(x['Чему вы научились в рамках курса?'], 
                                                                               'NOUN'), axis=1)

# подсчет глаголов (личная форма)
main_text['verbs'] = main_text.apply(lambda x: count_pos(x['value'], 'VERB'), axis=1)
additional_data_text['verbs'] = additional_data_text.apply(lambda x: count_pos(x['Чему вы научились в рамках курса?'], 
                                                                               'VERB'), axis=1)

# подсчет глаголов (инфинитив)
main_text['inf'] = main_text.apply(lambda x: count_pos(x['value'], 'INFN'), axis=1)
additional_data_text['inf'] = additional_data_text.apply(lambda x: count_pos(x['Чему вы научились в рамках курса?'], 
                                                                               'INFN'), axis=1)

# подсчет прилагательных
main_text['adj'] = main_text.apply(lambda x: count_pos(x['value'], 'ADJF'), axis=1)
additional_data_text['adj'] = additional_data_text.apply(lambda x: count_pos(x['Чему вы научились в рамках курса?'], 
                                                                               'ADJF'), axis=1)

# подсчет наречий
main_text['advb'] = main_text.apply(lambda x: count_pos(x['value'], 'ADVB'), axis=1)
additional_data_text['advb'] = additional_data_text.apply(lambda x: count_pos(x['Чему вы научились в рамках курса?'], 
                                                                               'ADVB'), axis=1)

# подсчет местоимений
main_text['pronoun'] = main_text.apply(lambda x: count_pos(x['value'], 'NPRO'), axis=1)
additional_data_text['pronoun'] = additional_data_text.apply(lambda x: count_pos(x['Чему вы научились в рамках курса?'], 
                                                                               'NPRO'), axis=1)

additional_data_text[['Чему вы научились в рамках курса?','nouns', 'verbs','inf','adj','advb','pronoun']]

### Features: кол-во слов с ошибками

Для проверки правописания русских и английских слов спользуем библиотеку `pyenchant`.

import re
import enchant
from enchant.checker import SpellChecker
from nltk.tokenize import ToktokTokenizer

# инициализируем словари
d_ru = enchant.DictWithPWL('ru')
d_eng = enchant.DictWithPWL('ru')

#инициализируем токенайзер
toktok = ToktokTokenizer()

def spell_check(line):
    char_regex = re.compile(r'[^а-яa-z]')
    line = char_regex.sub(' ', line.lower())
    
    chkr_ru = SpellChecker(d_ru)
    chkr_ru.set_text(line)
    mistakes_ru = set([err.word for err in chkr_ru])
    
    chkr_eng = SpellChecker(d_eng)
    chkr_eng.set_text(line)
    mistakes_eng = set([err.word for err in chkr_eng])
    
    tokenized = toktok.tokenize(line)
    num_mistakes=0
    for i in tokenized:
        if i in mistakes_ru and i in mistakes_eng:
            num_mistakes+=1
            
    return num_mistakes

main_text['num_mistakes'] = main_text['value'].apply(spell_check)

additional_data_text['num_mistakes'] = additional_data_text['Чему вы научились в рамках курса?'].apply(spell_check)

main_text['num_mistakes'].value_counts()

additional_data_text['num_mistakes'].value_counts()

### Агрегация средних значений выделенных признаков по каждому юзеру

agg_additional_data_text = additional_data_text[['untiID', 'sentiment_negative', 'sentiment_neutral','sentiment_positive', 
                                                 'sentiment_skip','sentiment_speech','ner','answer_len',
                                                 'upper_case_word_count','punctuation_count',
                                                'nouns', 'verbs','inf','adj','advb','pronoun',
                                                 'num_mistakes']].groupby(['untiID'],
                                                      as_index=False).agg(np.mean)

agg_main_text = main_text[['untiID', 'sentiment_negative', 'sentiment_neutral','sentiment_positive', 
                           'sentiment_skip','sentiment_speech','ner','answer_len',
                           'upper_case_word_count','punctuation_count',
                          'nouns', 'verbs','inf','adj','advb','pronoun','num_mistakes']].groupby(['untiID'],
                       as_index=False).agg(np.mean)


agg_additional_data_text.head(1)

agg_main_text.head(1)

# Объединим основные и доп агрегаты
agg = pd.concat([agg_main_text, agg_additional_data_text])
agg.shape

### Финальные датасеты
#### Объединим агрегированные текст.фичи и агрегаты диагностик 

df_train = pd.read_csv('diag_large_train.csv')
df_test = pd.read_csv('diag_test_agg.csv')

df_train.shape, df_test.shape

train = pd.merge(agg, df_train, on=['untiID'])
test = pd.merge(agg, df_test, on=['untiID'])
test = test[train.columns.tolist()]
train.shape, test.shape

train.to_csv('train.csv',index=False)
test.to_csv('test.csv',index=False)
