==> 5_Correlations.ipynb <==
### План:
- Заполнить пропуски в данных диагностики
- Агрегировать результаты диагностики по каждому юзеру
- Выделить фичи из текста
- Агрегировать фичи по юзеру
- Объединить датасеты
- __Построить матрицу корреляций по полученному датасету__
- __Наиболее удачные, яркие корреляции рассмотреть подробнее__
- Построить бейзлайн, лр/бустинг, для прогнозирования результата диагностики по фичам текста

import os
import numpy as np
import pandas as pd
import seaborn as sns

import matplotlib.pyplot as plt
%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

train = pd.read_csv('train.csv')
test = pd.read_csv('test.csv')

train.shape, test.shape

# определяем синтетические фичи
synth = ['sentiment_negative', 'sentiment_neutral','sentiment_positive', 'sentiment_skip', 
         'sentiment_speech','answer_len', 'upper_case_word_count', 'punctuation_count', 'ner']

# наименования архетипов
types = ['актер_sum','адепт_sum','аналитик_sum','боец_sum','визионер_sum','вождь_sum','интуит_sum','конструктор_sum',
'контролер_sum','эксперт_sum','организатор_sum','исполнитель_sum','преследователь_sum','экспериментатор_sum','связной_sum',
'творец_sum','хакер_sum','хозяин_sum','эмпат_sum','эстет_sum']

# Культура организационной деятельности
kod = ['автоматизация_sum','единая идеология компании_sum','поощрение инаковости_sum','исследование неопределённости_sum',
'поощрение работы над ошибками_sum','нестандартные решения_sum','гибкий график, удалённая работа_sum',
'мало гибких правил_sum','открытая система управления знаниями_sum','кооперация_sum','откровенность_sum',
'демократия_sum','краткосрочное планирование_sum','эффективность_sum','инновации_sum','вовлечённость_sum',
'динамичные команды_sum','обучение желающих_sum','власть экспертизы_sum','коллективизм_sum']

motiv = ['достижение, стремление к лидерству_mean','самосовершенствование, работа над собой, труд_mean',
        'социальный элемент, дружба, влияние, конкуренция_mean', 'осмысленность деятельности, собственная значимость_mean',
        'преодоление сложностей, препятствий_mean','избегание негатива, размеренность, безопасность_mean']

### Матрицы корреляций

Явные корреляции наблюдаются между некоторыми признаками из диагностики. Например:
- Архетип __"Аналитик"__ имеет положительную корреляцию с такими мотиваторами: _"осмысленность деятельности, собственная значимость"_, _"достижение, стремление к лидерству"_, _"самосовершенствование, работа над собой, труд"_.


- Архетип __"Вождь"__ имеют положительную корреляцию с мотиватором _"достижение, стремление к лидерству"_.



def corrFilter(x: pd.DataFrame, bound: float):
    xCorr = x.corr()
    xFiltered = xCorr[((xCorr >= bound) | (xCorr <= -bound)) & (xCorr !=1.000)]
    xFlattened = xFiltered.unstack().sort_values().drop_duplicates()
    xFlattened = xFlattened.reset_index()
    xFlattened.columns = ['col1','col2','value']
    return xFlattened

corrFilter(train[types+motiv], 0.55)

Также признаки, извлеченные из текстов, коррелируют между собой. Например, есть положительная корреляция между _"Длиной сообщения"_ и _"Количеством знаков препинания"_.

corrFilter(train[synth+types], 0.55)

Корреляций между выделенными текстовыми признаками и значениями диагностик мало.

corrFilter(train[synth+types+kod+motiv], 0.55)
